Build Windows V1 and V2 runtimes always
