+++ services/tasks/local_unix.go
@@ -1,5 +1,3 @@
-// +build !windows_v2
-
 /*
    Copyright The containerd Authors.
